Add stubs for GraphQL endpoints
