+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/auth/webservice/authGraphQLParams.kt
@@ -2,6 +2,8 @@ package app.ehrenamtskarte.backend.auth.webservice
 
 import app.ehrenamtskarte.backend.auth.webservice.dataloader.ADMINISTRATOR_LOADER_NAME
 import app.ehrenamtskarte.backend.auth.webservice.dataloader.administratorLoader
+import app.ehrenamtskarte.backend.auth.webservice.schema.ChangePasswordMutationService
+import app.ehrenamtskarte.backend.auth.webservice.schema.ResetPasswordMutationService
 import app.ehrenamtskarte.backend.auth.webservice.schema.SignInMutationService
 import app.ehrenamtskarte.backend.common.webservice.GraphQLParams
 import com.expediagroup.graphql.generator.SchemaGeneratorConfig
@@ -18,7 +20,9 @@ val authGraphQlParams = GraphQLParams(
     config = SchemaGeneratorConfig(supportedPackages = listOf("app.ehrenamtskarte.backend.auth.webservice.schema")),
     dataLoaderRegistry = createDataLoaderRegistry(),
     mutations = listOf(
-        TopLevelObject(SignInMutationService())
+        TopLevelObject(SignInMutationService()),
+        TopLevelObject(ChangePasswordMutationService()),
+        TopLevelObject(ResetPasswordMutationService())
     ),
     queries = listOf()
 )
